+++ tests/sim/status_warehouse/container/carousel/test_carousel.py
@@ -10,7 +10,7 @@ class TestCarousel(TestCase):
     def setUp(self):
         self.warehouse = Warehouse()
         self.carousel_config = CarouselConfiguration(
-            length=200,
+            length=450,
             bay_height = 150,
             buffer_height = 150,
             x_offset = 125,